state: Moved machine-specified interfaces handling to a separate file
